XMLHttpRequest/responsetype.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>XMLHttpRequest.responseType</title>
<link rel="author" title="Mathias Bynens" href="http://mathiasbynens.be/">
<link rel="author" title="Ms2ger" href="mailto:Ms2ger@gmail.com">
<link rel="help" href="http://xhr.spec.whatwg.org/#the-responsetype-attribute">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id="log"></div>
<script>
var types = ['', 'json', 'document', 'arraybuffer', 'blob', 'text'];
types.forEach(function(type) {
  test(function() {
    var xhr = new XMLHttpRequest();
    xhr.open('get', '/');
    xhr.responseType = type;
    assert_equals(xhr.responseType, type);
  }, 'Set responseType to ' + format_value(type) + ' when readyState is OPENED.');
});
</script>
